+++ cgateweb/src/cgateWebBridge.js
@@ -165,6 +165,7 @@ class CgateWebBridge {
             allowUnauthenticatedMutations: this.settings.web_allow_unauthenticated_mutations === true,
             allowedOrigins: this.settings.web_allowed_origins || null,
             maxMutationRequestsPerWindow: this.settings.web_mutation_rate_limit_per_minute || 120,
+            triggerAppId: this.settings.ha_discovery_trigger_app_id || null,
             getStatus: () => this._getBridgeStatus()
         });
         this.haBridgeDiagnostics = new HaBridgeDiagnostics(